requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
Django==3.1.1
django-appconf==1.0.4
django-ckeditor==6.0.0
django-ckeditor-filebrowser-filer==0.3.0
django-classy-tags==2.0.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-uuslug==1.2.0
idna==2.10
importlib-metadata==1.7.0
pilkit==2.0
Pillow==7.2.0
pytest-runner==5.2
python-slugify==4.0.1
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Unidecode==1.1.1
urllib3==1.26.2
zipp==3.1.0